Change package layout
* Remove types/
* Introduce CommentSource and source/ packages
