+++ provider/mistral.go
@@ -3,7 +3,7 @@ package provider
 import (
 	"fmt"
 
-	summaraizer "github.com/ioki-mobility/summaraizer/types"
+	"github.com/ioki-mobility/summaraizer"
 )
 
 type Mistral struct {
@@ -11,7 +11,7 @@ type Mistral struct {
 	ApiToken string
 }
 
-func (m Mistral) Summarize(
+func (m *Mistral) Summarize(
 	comments summaraizer.Comments,
 ) (string, error) {
 	return fmt.Sprintf("This is a summary from mistral with model %s", m.Model), nil
