+++ provider/openai.go
@@ -3,7 +3,7 @@ package provider
 import (
 	"fmt"
 
-	summaraizer "github.com/ioki-mobility/summaraizer/types"
+	"github.com/ioki-mobility/summaraizer"
 )
 
 type OpenAi struct {
@@ -11,7 +11,7 @@ type OpenAi struct {
 	ApiToken string
 }
 
-func (o OpenAi) Summarize(
+func (o *OpenAi) Summarize(
 	comments summaraizer.Comments,
 ) (string, error) {
 	return fmt.Sprintf("This is a summary from OpenAI with model %s", o.Model), nil
