--- types/types.go
@@ -1,8 +1,0 @@
-package types
-
-type Comments = []Comment
-
-type Comment struct {
-	Author string
-	Body   string
-}
